streamlit-csv-ai-app/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openai>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
python-dotenv>=1.0.0
requests>=2.0.0
mysql-connector-python>=8.0.0